Modified to addd autoid

+++ productdata/src/main/java/com/surendra/springdata/product/entities/Product.java
@@ -1,23 +1,27 @@
 package com.surendra.springdata.product.entities;
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+
 import javax.persistence.Column;
 @Entity
 @Table
 public class Product {
 	
 	@Id
-	private int id;
+	//@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 	private String name;
 	@Column(name = "description")
 	private String desc;
 	private Double price;
 	
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
